super commit

+++ src/main/java/br/ufrn/bti/banco1000/model/SalaryAccount.java
@@ -1,5 +1,7 @@
 package br.ufrn.bti.banco1000.model;
 
+import java.math.BigDecimal;
+
 import br.ufrn.bti.banco1000.model.enumerations.AccountType;
 
 public class SalaryAccount extends Account {
@@ -7,7 +9,7 @@ public class SalaryAccount extends Account {
     private final int withdrawalLimit = 2;
     
     public SalaryAccount(int agency, Long accountNumber, Long clientId, AccountType accountType, int password,
-            double balance) {
+            BigDecimal balance) {
         super(agency, accountNumber, clientId, accountType, password, balance);
     }
